# lib/api-client-react/src/generated.schemas.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Api
 * ECU Vehicle Monitoring API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthCredentials {
  email: string;
  password: string;
}

export interface User {
  id: string;
  email: string;
  createdAt: string;
}

export interface EcuSensorInput {
  /** Vehicle speed in km/h */
  speed: number;
  /** Engine RPM */
  rpm: number;
  /** Engine temperature in Celsius */
  engineTemp: number;
  /** Fuel level percentage 0-100 */
  fuelLevel: number;
  /** Battery voltage in volts */
  batteryVoltage: number;
  /** Throttle position percentage */
  throttlePosition?: number;
  /** Coolant temperature in Celsius */
  coolantTemp?: number;
  /** Oil pressure in PSI */
  oilPressure?: number;
}

export interface EcuReading {
  id: string;
  speed: number;
  rpm: number;
  engineTemp: number;
  fuelLevel: number;
  batteryVoltage: number;
  throttlePosition: number;
  coolantTemp: number;
  oilPressure: number;
  hasFault: boolean;
  /** simulation or esp32 or external */
  source: string;
  createdAt: string;
}

export interface DiagnosticTroubleCode {
  id: string;
  /** DTC code e.g. P0101 */
  code: string;
  description: string;
  /** critical, warning, info */
  severity: string;
  /** engine, fuel, electrical, cooling, transmission */
  system: string;
  isActive: boolean;
  detectedAt: string;
  /** @nullable */
  resolvedAt: string | null;
  /** @nullable */
  ecuReadingId: string | null;
}

export interface VehicleStatus {
  /** healthy, warning, critical */
  overallHealth: string;
  activeFaultCount: number;
  criticalFaultCount: number;
  warningFaultCount: number;
  totalReadings: number;
  /** @nullable */
  lastReadingAt: string | null;
  latestReading: EcuReading;
}

export interface HistoryStats {
  speed: MetricStats;
  rpm: MetricStats;
  engineTemp: MetricStats;
  fuelLevel: MetricStats;
  batteryVoltage: MetricStats;
}

export interface MetricStats {
  min: number;
  max: number;
  avg: number;
  current: number;
}

export interface ErrorResponse {
  error: string;
}

export type Register200 = {
  message?: string;
};

export type ListDtcsParams = {
/**
 * Filter by active status
 */
active?: boolean;
/**
 * Max records to return
 */
limit?: number;
};

export type GetHistoryParams = {
/**
 * Number of records (default 50)
 */
limit?: number;
/**
 * Offset for pagination
 */
offset?: number;
};